+++ src/main/java/cart/common/GlobalExceptionHandler.java
@@ -1,5 +1,7 @@
 package cart.common;
 
+import cart.common.auth.AuthenticationException;
+import cart.dto.ExceptionDto;
 import java.util.Objects;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
