+++ src/main/java/cart/config/WebConfig.java
@@ -1,7 +1,7 @@
 package cart.config;
 
-import cart.common.AuthInfoHandlerMethodArgumentResolver;
-import cart.common.AuthenticationCheckInterceptor;
+import cart.common.auth.AuthInfoHandlerMethodArgumentResolver;
+import cart.common.auth.AuthenticationCheckInterceptor;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
