+++ src/main/java/cart/domain/auth/service/AuthService.java
@@ -1,6 +1,6 @@
 package cart.domain.auth.service;
 
-import cart.common.AuthenticationException;
+import cart.common.auth.AuthenticationException;
 import cart.domain.member.dao.MemberDao;
 import cart.domain.member.entity.Member;
 import cart.dto.MemberInformation;
